sync Swift data with minimal role assignments
Swift allows the reseller admin to operate on all projects while the
reseller admin role is assigned to him only on one project. So now we
assign the reseller admin role only on the service user project, and use
the connection to that project to retrieve quota/usage data for all
projects.

Closes #89.