work around a rendering glitch in Firefox
Most of the time, the div.progress-bar will contain one or multiple
div.progress (with display: inline-block) whose width adds up to 100%.
They therefore should be rendered on one line, as expected for a
progress bar.

However, likely due to rounding errors in the rendering engine,
sometimes the calculated widths add up to > 100%, causing the last
div.progress to be moved to the next line. (Since there is no room for a
second line in a progress bar, this means that the last div.progress
becomes invisible.)

To remedy that, we subtract 0.1 width-% from the rightmost bar. This is
less than one pixel, so technically invisible, but circumvents the
rounding errors and thus fixes the display in the cases that I have
observed.