unbreak file uploading temporarily
For some reason, when the initial upload form is rendered,
`request.xhr?` evaluates to false even though the form was clearly
`X-Requested-With: XMLHttpRequest`. The "Save" button on the form
therefore does a regular POST (rather than an AJAXy one) that accepts
only HTML, so the JS template cannot be rendered.

As a workaround, enable `format.html` to navigate back to the
object listing where the uploaded object will then be visible.