add a simple handling for backend errors
The whole error reporting situation in the Core::ServiceLayer is kind of
a mess, so this mostly works around it for now, and reads plain Excon
errors.

That's enough for now since Swift actually returns useful status codes.
The response body seems not to add much to the status code, so we just
ignore it for now.