fix metric collection
The latest ruby prometheus client needs the label builds defined slightly different.

Also the latency metric is now a histogram instead of a summary. The prometheus graphs need to be changed to work again.

Also I reduced the label cardianlity on the latency metrics, now we only record the request method and the plugin.