Adds Automatic Changelog Generation
This commits adds a Docker image that includes a tool that automatically
generates changelogs.

See: https://github.com/skywinder/github-changelog-generator

This tool relies on Github metadata to generate the Changelog. This has
the advantage that it knows more than the Git history e.g. PRs,
Labels...

The here proposed workflow is to not include the generation of the
changelog into CI, but rather manually run it for each major version.
It's easy enough:

  rm CHANGELOG.md
  make CHANGELOG.md