Fixed changelog generator
The include_body_issue PR was rebased to master before merged in the original repo. Since beginning of the Fork the master branch have been commited and also a bug introduce which i reported in this issue: https://github.com/github-changelog-generator/github-changelog-generator/issues/651

Until this is not being fixed, we should stay in the fork.

 [ci skip]