Policymod secgroups (#471)
* Added two new roles for securitygroup administration and validation

* Added new roles securitygroup_viewer and securitygroup_admin to the
policy file. In addition the policy file is a bit simplyfied with merging compute_admin and network_admin
into the context securitygroup_admin and assigned the permission for security groups
only to this context. Same done with the *viewer permissions.

* Added option for securitygroup_admin to
add, edit and delete securitygroup rules and for the securitygroup_viewer to list
and show them.