[lbaas] Fix wording: Members are marked as OFFLINE
Customers rarely use the dashboard and the CLI at the same time, so
it's better to keep the wording true to the concepts used on the dashboard.

In #944 we decided to display members in state ERROR as OFFLINE.
It is true that the operating_status of the members is set to ERROR in
Octavia, but it is marked as OFFLINE in Elektra.